Fix more clang warnings
* libwget/css_tokenizer.lex: Skip -Wsuggest-attribute=pure for clang
* libwget/iri.c (wget_iri_parse): Explicit cast
* src/dl.c: Use 'int' for vector size
* src/options.c: Likewise,
  remove 'return' after call to exit(),
  (set_long_option): Rename local var 'env' to 'path'
* src/plugin.c: Make global vars 'static',
  Use 'int' for vector size

+++ libwget/iri.c
@@ -470,7 +470,7 @@ wget_iri_t *wget_iri_parse(const char *url, const char *encoding)
 			if (c_isdigit(s[1])) {
 				int port = atoi(s + 1);
 				if (port > 0 && port < 65536) {
-					iri->port = port;
+					iri->port = (uint16_t) port;
 					iri->port_given = true;
 				}
 			}
